all: rewrite multiple pkgs; streamline demo; etc
- Add -v option to got-reload, which is passed to "go run".
- Split the demo example package into two packages, example and
  example2.
- Make the reloader work with multiple packages.
- Make the reloader use a private instance of log.Logger.
- "got-reload run" adds a grl_dependencies.go file to each package being
  rewritten, and then deletes them when done.  (Unless interrupted by
  SIGINT/^C).
- The target program doesn't need to explicitly start the reloader and
  more.
- The rewritten packages don't need to import reflect or gotreload any
  more.

These last two are achieved via the grl_dependencies.go file mentioned
earlier.  It's a constant (except for package name) and looks like this:

  package %s

  import (
	  "reflect"

	  "github.com/huckridgesw/got-reload/pkg/gotreload"
	  _ "github.com/huckridgesw/got-reload/pkg/reloader/start"
  )

  var (
	  _ = reflect.ValueOf
	  _ = gotreload.RegisterAll
  )

Every rewritten package gets this added to it.  The imports and
assignments make the Register calls in the generated code work, and the
import of "github.com/huckridgesw/got-reload/pkg/reloader/start" has an
init() function that starts the reloader, if $GOT_RELOAD_START_RELOADER
is set to 1.

Signed-off-by: Larry Clapp <larry@theclapp.org>